Separated express app and all config in separate file and formatting
